Materialize it
I thought it would be nice to have nicer styling for the inputs with materialize. To make this work I had to make some modifications to the way things were done. For example, checkboxes have to be setup a certain way with materialize to actually display.

Also, I started breaking up logic.js into smaller pieces. The larger file is really hard to reason about.