Apply integration between the deck list and the vanilla js app
I'm trying to do this migration in as small of chunks as possible. The 'apply' step seemed like a good place to break off for this piece of functionality. To make this work I just had to change the click handler for the old apply button and instead add a similar method to the window object. The DeckList will call indo this with the correct selection data when the apply button is clicked.