Drop master deck-list import
Deck list is now no longer needed except as part of the init of the vanilla js app. Currently it is still being used in the scenario list. This likely means the scenario list is broken at the moment, so I may need to check that out at some point.