Make decks extensible
I want to make it easy to plug in alternative data sources without mashing the data together. To do this I'm going to make the card service merge the data. I'm also going to include group headers for the deck objects so its easy to scan through and see where a group ends and the next one begins. I might consider doing this for the gloomhaven decks too.