Shorten the monster stats storage format
I want to make it simpler to create new monsters without having to fill out so much data. To do this I created a new format where you specify the base stats of the unit, and then a list of upgrades for each level (for normal and elite, plus one 'base' upgrade for level 0 elites). The monster service then 'denormalizes' the data by marking up the monster stats by the upgrade amount for each level to populate the complete list of levels.

This shortens the format by only requiring you to specify what changes at each level. Note that if there are any changes to the attributes, you will need to respecify the full list of attributes. It would be too difficult to figure out how to splice the upgrades together with the previous attributes.